Move the interface's startup into a library, unchanged
Pure relocation, so that the commit after it shows only what actually changes.
Everything `main` did apart from installing the logger becomes
`mirrord_tui::run`, lines untouched; `main.rs` keeps the logger and calls it.

Four lines are not carried over verbatim, and they are the whole of it: the
`use std::{fs::File, io::stdout, sync::Arc}` line splits across the two files,
`async fn main` becomes `pub async fn run`, and `main` gains the call to it.

One thing does reorder. The logger used to be installed after the first
`execute!`, and is now installed before `run` is called at all, because the two
ended up in different files. Nothing is drawn either way, so nothing observable
moves with it.
